Move logger test to source tree
